Update to protobuf3 and fix compiler warnings

Approved-by: Carlos Agüero <cen.aguero@gmail.com>
Approved-by: Steven Peters <scpeters@osrfoundation.org>

+++ src/Node_TEST.cc
@@ -103,7 +103,7 @@ void cbInfo(const ignition::msgs::Int32 &_msg,
 }
 
 //////////////////////////////////////////////////
-void rawCbInfo(const char *_msgData, const int _size,
+void rawCbInfo(const char *_msgData, const size_t _size,
                const ignition::transport::MessageInfo &_info)
 {
   EXPECT_EQ(_info.Topic(), g_topic);
